chore: remove unused media lib files
