+++ packages/core/upload/admin/src/pages/App/MediaLibrary.tsx
@@ -25,27 +25,27 @@ import { useIntl } from 'react-intl';
 import { Link as ReactRouterLink, useNavigate, useLocation } from 'react-router-dom';
 import { styled } from 'styled-components';
 
-import { AIUploadModal } from '../../../ai/components/AIUploadModal';
-import { AssetGridList } from '../../../components/AssetGridList/AssetGridList';
-import { EditAssetDialog } from '../../../components/EditAssetDialog/EditAssetContent';
-import { EditFolderDialog } from '../../../components/EditFolderDialog/EditFolderDialog';
-import { FolderCard } from '../../../components/FolderCard/FolderCard/FolderCard';
-import { FolderCardBody } from '../../../components/FolderCard/FolderCardBody/FolderCardBody';
-import { FolderCardBodyAction } from '../../../components/FolderCard/FolderCardBodyAction/FolderCardBodyAction';
-import { FolderCardCheckbox } from '../../../components/FolderCard/FolderCardCheckbox/FolderCardCheckbox';
-import { FolderGridList } from '../../../components/FolderGridList/FolderGridList';
-import { SortPicker } from '../../../components/SortPicker/SortPicker';
-import { TableList } from '../../../components/TableList/TableList';
-import { UploadAssetDialog } from '../../../components/UploadAssetDialog/UploadAssetDialog';
-import { localStorageKeys, viewOptions } from '../../../constants';
-import { useAIAvailability } from '../../../hooks/useAiAvailability';
-import { useAssets } from '../../../hooks/useAssets';
-import { useFolder } from '../../../hooks/useFolder';
-import { useFolders } from '../../../hooks/useFolders';
-import { useMediaLibraryPermissions } from '../../../hooks/useMediaLibraryPermissions';
-import { usePersistentState } from '../../../hooks/usePersistentState';
-import { useSelectionState } from '../../../hooks/useSelectionState';
-import { containsAssetFilter, getBreadcrumbDataML, getFolderURL, getTrad } from '../../../utils';
+import { AIUploadModal } from '../../ai/components/AIUploadModal';
+import { AssetGridList } from '../../components/AssetGridList/AssetGridList';
+import { EditAssetDialog } from '../../components/EditAssetDialog/EditAssetContent';
+import { EditFolderDialog } from '../../components/EditFolderDialog/EditFolderDialog';
+import { FolderCard } from '../../components/FolderCard/FolderCard/FolderCard';
+import { FolderCardBody } from '../../components/FolderCard/FolderCardBody/FolderCardBody';
+import { FolderCardBodyAction } from '../../components/FolderCard/FolderCardBodyAction/FolderCardBodyAction';
+import { FolderCardCheckbox } from '../../components/FolderCard/FolderCardCheckbox/FolderCardCheckbox';
+import { FolderGridList } from '../../components/FolderGridList/FolderGridList';
+import { SortPicker } from '../../components/SortPicker/SortPicker';
+import { TableList } from '../../components/TableList/TableList';
+import { UploadAssetDialog } from '../../components/UploadAssetDialog/UploadAssetDialog';
+import { localStorageKeys, viewOptions } from '../../constants';
+import { useAIAvailability } from '../../hooks/useAiAvailability';
+import { useAssets } from '../../hooks/useAssets';
+import { useFolder } from '../../hooks/useFolder';
+import { useFolders } from '../../hooks/useFolders';
+import { useMediaLibraryPermissions } from '../../hooks/useMediaLibraryPermissions';
+import { usePersistentState } from '../../hooks/usePersistentState';
+import { useSelectionState } from '../../hooks/useSelectionState';
+import { containsAssetFilter, getBreadcrumbDataML, getFolderURL, getTrad } from '../../utils';
 
 import { BulkActions } from './components/BulkActions';
 import { EmptyOrNoPermissions } from './components/EmptyOrNoPermissions';
@@ -54,11 +54,11 @@ import { Header } from './components/Header';
 
 import type { BulkActionsProps } from './components/BulkActions';
 import type { HeaderProps } from './components/Header';
-import type { Query } from '../../../../../shared/contracts/files';
-import type { FolderDefinition } from '../../../../../shared/contracts/folders';
-import type { AssetGridListProps } from '../../../components/AssetGridList/AssetGridList';
-import type { Asset } from '../../../components/EditAssetDialog/EditAssetContent';
-import type { FolderRow, FileRow, TableListProps } from '../../../components/TableList/TableList';
+import type { Query } from '../../../../shared/contracts/files';
+import type { FolderDefinition } from '../../../../shared/contracts/folders';
+import type { AssetGridListProps } from '../../components/AssetGridList/AssetGridList';
+import type { Asset } from '../../components/EditAssetDialog/EditAssetContent';
+import type { FolderRow, FileRow, TableListProps } from '../../components/TableList/TableList';
 
 const BoxWithHeight = styled(Box)`
   height: 3.2rem;
