+++ packages/core/upload/admin/src/pages/App/components/BulkActions.tsx
@@ -11,6 +11,7 @@ import type {
   FolderDefinition,
   Folder as FolderInitial,
 } from '../../../../../shared/contracts/folders';
+
 interface FolderWithType extends FolderInitial {
   type: string;
 }
@@ -19,7 +20,7 @@ export interface FileWithType extends File {
   type: string;
 }
 
-interface BulkActionsProps {
+export interface BulkActionsProps {
   selected: Array<FileWithType | FolderDefinition> | Array<FolderWithType | FileWithType>;
   onSuccess: () => void;
   currentFolder?: FolderWithType;
@@ -27,6 +28,11 @@ interface BulkActionsProps {
 
 export const BulkActions = ({ selected = [], onSuccess, currentFolder }: BulkActionsProps) => {
   const { formatMessage } = useIntl();
+  const numberAssets = selected?.reduce(function (_this, val) {
+    return val?.type === 'folder' && 'files' in val && val?.files && 'count' in val.files
+      ? _this + val?.files?.count
+      : _this + 1;
+  }, 0);
 
   return (
     <Flex gap={2} paddingBottom={5}>
@@ -38,8 +44,8 @@ export const BulkActions = ({ selected = [], onSuccess, currentFolder }: BulkAct
               '{numberFolders, plural, one {1 folder} other {# folders}} - {numberAssets, plural, one {1 asset} other {# assets}} selected',
           },
           {
-            numberFolders: selected.filter(({ type }) => type === 'folder').length,
-            numberAssets: selected.filter(({ type }) => type === 'asset').length,
+            numberFolders: selected?.filter(({ type }) => type === 'folder').length,
+            numberAssets,
           }
         )}
       </Typography>
