+++ tests/scripts/run-cli-tests.js
@@ -4,4 +4,5 @@
 // Inject --type cli before the rest of the arguments
 const args = process.argv.slice(2);
 process.argv = [process.argv[0], process.argv[1], '--type', 'cli', ...args];
+// eslint-disable-next-line import/extensions
 require('./run-tests.js');
